+++ src/Entities/Traits/EntityTrait.php
@@ -11,7 +11,7 @@ namespace League\OAuth2\Server\Entities\Traits;
 
 trait EntityTrait
 {
-    /*
+    /**
      * @var string
      */
     protected $identifier;