+++ src/Entities/Traits/TokenEntityTrait.php
@@ -25,7 +25,7 @@ trait TokenEntityTrait
     protected $expiryDateTime;
 
     /**
-     * @var string|int
+     * @var string|int|null
      */
     protected $userIdentifier;
 
@@ -89,7 +89,7 @@ trait TokenEntityTrait
     /**
      * Set the identifier of the user associated with the token.
      *
-     * @param string|int $identifier The identifier of the user
+     * @param string|int|null $identifier The identifier of the user
      */
     public function setUserIdentifier($identifier)
     {
@@ -99,7 +99,7 @@ trait TokenEntityTrait
     /**
      * Get the token user's identifier.
      *
-     * @return string|int
+     * @return string|int|null
      */
     public function getUserIdentifier()
     {