+++ tests/CryptKeyTest.php
@@ -3,8 +3,9 @@
 namespace LeagueTests\Utils;
 
 use League\OAuth2\Server\CryptKey;
+use PHPUnit\Framework\TestCase;
 
-class CryptKeyTest extends \PHPUnit_Framework_TestCase
+class CryptKeyTest extends TestCase
 {
     /**
      * @expectedException \LogicException
@@ -26,6 +27,14 @@ class CryptKeyTest extends \PHPUnit_Framework_TestCase
     public function testKeyFileCreation()
     {
         $keyContent = file_get_contents(__DIR__ . '/Stubs/public.key');
+        $key = new CryptKey($keyContent);
+
+        $this->assertEquals(
+            'file://' . sys_get_temp_dir() . '/' . sha1($keyContent) . '.key',
+            $key->getKeyPath()
+        );
+
+        $keyContent = file_get_contents(__DIR__ . '/Stubs/private.key.crlf');
         $key = new CryptKey($keyContent);
 
         $this->assertEquals(