+++ tests/CryptTraitTest.php
@@ -2,10 +2,10 @@
 
 namespace LeagueTests\Utils;
 
-use League\OAuth2\Server\CryptKey;
 use LeagueTests\Stubs\CryptTraitStub;
+use PHPUnit\Framework\TestCase;
 
-class CryptTraitTest extends \PHPUnit_Framework_TestCase
+class CryptTraitTest extends TestCase
 {
     /**
      * @var \LeagueTests\Stubs\CryptTraitStub